requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
PyJWT==2.8.0
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
django-channels-jwt-auth-middleware==1.0.0